// ws-client/src/app/wild-snap/common/shared/model/textbook/textbook-list.ts
export class TextbookList {

  oid: string;
  nameEn: string;
  nameBn: string;
  subjectCode: string;
  eBookLink: string;
  status: string;
  instituteOid: string;
  instituteSessionOid: string;
  instituteVersionOid: string;
  instituteClassGroupOid: string;
  instituteClassOid: string;
  educationTextbookOid: string;
  noOfStudent: number;

  value: boolean;

  //used in service
}